+++ x/rewards/ante/ante_utils.go
@@ -11,7 +11,7 @@ import (
 // RewardsKeeperExpected defines the expected interface for the x/rewards keeper.
 type RewardsKeeperExpected interface {
 	// Used in MinFeeDecorator
-	GetMinConsensusFee(ctx sdk.Context) (sdk.DecCoin, bool)
+	ComputationalPriceOfGas(ctx sdk.Context) sdk.DecCoin
 	GetFlatFee(ctx sdk.Context, contractAddr sdk.AccAddress) (sdk.Coin, bool)
 	CreateFlatFeeRewardsRecords(ctx sdk.Context, contractAddress sdk.AccAddress, flatfee sdk.Coins)